+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 Contributed:
 
-- Improve `rpc-provider` README (thanks to https://github.com/josepot)
+- Improve `rpc-provider` README (Thanks to https://github.com/josepot)
+- Add handling for `Vec<[<Type>;<length>]>` in typegen (Thanks to https://github.com/zannis)
 
 Changes: